Moved glob code from filesystem.dart to preferences.dart
